TGSH_GridPlate Holder_A
'TGSH_GridPlate Holder_A' is a piece that needs to be 3D printed, filled with steel washers and glued to piece 'TGSH_GridPlate Holder_B'.  The TGSH_GridPlate Holder assembly is designed to be wrapped with stretched Parafilm to hold a single Hiraoka Gridplate that pinches and holds up to 20 TEM grids for post-grid uranyl acetate and Sato's lead contrast staining.