Add files via upload
'TGSH_GridPlate Holder_B' is a piece that needs to be 3D printed, glued to piece 'TGSH_GridPlate Holder_A'.  The TGSH_GridPlate Holder assembly is designed to be wrapped with stretched Parafilm to hold a single Hiraoka Gridplate that pinches and holds up to 20 TEM grids for post-grid uranyl acetate and Sato's lead contrast staining.